- fixed tune() - memory reading/writing changed to new system

+++ avr/transceiver.h
@@ -8,7 +8,7 @@ void set_modulation(unsigned int mod);
 int rogerbeep(void);
 void off(void);
 void channel(unsigned int ch);
-int ch2freq(unsigned int ch);
+unsigned long ch2freq(unsigned int ch);
 void set_ctcss(unsigned int ctcss_value);
 void set_rpt(unsigned int rpt_value);
 void set_echo(unsigned int echo_value);
